requirements.txt
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.9.4
drf-nested-routers==0.91